Fix double enable errors
If 'enable' was called on an existing instance that was mid-drag then
the old drag would get stuck and throw errors. This fixes that by
cleaning up the old instance and only performing actions on move or
dragend when this instance started the drag.

Bumped to v2.0.0 due to dropping oldIE support.

 - Add checks to ensure that we have started dragging before handling a
   move
 - Add removal of prior $placeholder and $ghost elements if a prior
   instance was actively dragging
  - This drops the previous drag at the spot it was at when the
    re-enable happened
 - Remove IE8 & IE9 support