Fix fixed position ghosts inside of transform elements
Instead of appending the `$ghost` element to the parent of the
container, we'll create a `.ghost-container` at the body DOM root and
append the `$ghost` element there.

Due to parent elements or containing parent elements with CSS
`transform` properties. When the parent container has the CSS
`transform` property, it "creates" a new coordinate system as if it's a
different viewport than the viewport you're viewing in. Because of this,
any `fixed` children elements will behave as if they are `absolute`ly
positioned rather than `fixed`.

This issue is only in webkit-based browsers. There is currently bug
filed here: https://bugs.chromium.org/p/chromium/issues/detail?id=20574.

With this "hack" the assumption is that `body` doesn't not have a CSS
`transform` property on it.